% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanCRC32-function.R
\name{clanCRC32}
\alias{clanCRC32}
\title{Version function for the clanCRC32 library}
\usage{
clanCRC32()
}
\value{
A list of information about the version and functions within clanCRC32.
}
\description{
Version function for the clanCRC32 library.
}
\author{
Claus E. Andersen
}
