% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-hex-function.R
\name{convert.hex}
\alias{convert.hex}
\title{Convert hex string to decimal number.}
\usage{
convert.hex("FF")
}
\arguments{
\item{x}{input string (a hex number)}
}
\value{
integer
}
\description{
# This function converts hex strings to integers
without breaking the 2^31 limit.
}
\author{
Claus E. Andersen
}
