% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRC32-validation-function.R
\name{CRC32.validation}
\alias{CRC32.validation}
\title{CRC32 validation of given file}
\usage{
CRC32.validation("C:\\data\\projects\\R\\work\\exp100830-cobalt-calibrations\\exp100914.txt")
}
\arguments{
\item{file.name}{e.g. "c:\\data\\ME30-data\\catest20006a.txt",}

\item{delimiter}{e.g. ";"}

\item{stop.if.fault}{(TRUE or FALSE)}
}
\value{
a list with xlim-values for a trellis plot.
}
\description{
Each line is terminated with a CRC32 code. This function
eturns if all lines in the file can be validated.
}
\author{
Claus E. Andersen
}
