% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-hex-function.R
\name{convert.hex}
\alias{convert.hex}
\title{Convert hex strings to integers (needed in CRC32 computations)
Does notbreak the 2^31 limit.}
\usage{
convert.hex(x)
}
\arguments{
\item{x}{}
}
\value{
CRC32 value
}
\description{
Convert hex strings to integers (needed in CRC32 computations)
Does notbreak the 2^31 limit.
}
